convert component to typescript
