+++ common/views/components/GifVideo/GifVideo.js
@@ -4,6 +4,7 @@ import debounce from 'lodash.debounce';
 import throttle from 'lodash.throttle';
 import { font, classNames } from '../../../utils/classnames';
 import { trackEvent } from '../../../utils/ga';
+// $FlowFixMe (tsx)
 import Tasl from '../Tasl/Tasl';
 // $FlowFixMe (tsx)
 import Caption from '../Caption/Caption';
